demo/requirements.txt
# Core
fastapi[standard]==0.118.3
uvicorn==0.38.0
gunicorn==24.0.0
pydantic[email]==2.12.5
pydantic-settings==2.9.1
python-multipart==0.0.22
python-dotenv==1.0.1

# Database
pymongo==4.10.1
# Celery 5.4 requires redis<6; redis 5.x has redis.asyncio for the FastAPI app
redis[hiredis]>=4.5.2,<6.0.0

# Vector DB
qdrant-client==1.16.2

# AI / LLM
openai==1.97.0
numpy==2.1.3

# Auth
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Templating
jinja2==3.1.6

# HTTP
requests==2.32.4
httpx==0.28.1

# Rate limiting
slowapi==0.1.9

# Document processing
PyPDF2==3.0.1
python-docx==1.2.0
Pillow==11.3.0

# Computer vision / biometrics
opencv-python==4.11.0.86
# face_recognition==1.3.0  # requires dlib+CMake; not compatible with Python 3.13

# Task queue
celery[redis]==5.4.0

# Observability
structlog==25.1.0
prometheus-client==0.22.1

# Security
python-magic==0.4.27

# Object storage
boto3==1.38.0

# Testing (dev only)
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-cov==6.1.1
locust==2.35.0
ruff==0.11.4
